yaml_preprocessor: add support for $include
This commit adds support for a special `$include` keyword in the YAML
preprocessor. This new keyword is similar to the existing `$eval` keyword
and allows building the `robot.model.yaml` up from multiple separate files.

$include can process both absolute filenames as well as relative filenames,
which are interpreted relative to the file currently being processed.

The included document is also processed by the YAML preprocessor, and so can
include nested `$include`s and `$eval`s.

This commit also updates the documentation and tests.